fix: changes and viewmodel added
